core/src: Move half traits to private header and add missing infinity trait
